+++ WalWal/Features/MissionUpload/MissionUploadPresenter/Implement/Views/Component/CustomTextView.swift
@@ -27,8 +27,9 @@ final class StyledTextInputView: UIView {
   public private(set) var textView = UnderlinedTextView(
     font: Fonts.KR.H6.B,
     textColor: Colors.gray600.color,
+    tintColor: Colors.walwalOrange.color,
     underLineColor: Colors.gray800.color,
-    tintColor: Colors.walwalOrange.color
+    numberOfLines: 4
   ).then {
     $0.backgroundColor = .clear
     $0.isEditable = true
@@ -40,6 +41,7 @@ final class StyledTextInputView: UIView {
     $0.font = Fonts.EN.Caption
     $0.textColor = Colors.white.color.withAlphaComponent(0.2)
     $0.text = "0/80"
+    $0.textAlignment = .right
   }
   
   // MARK: - Properties
@@ -87,7 +89,7 @@ final class StyledTextInputView: UIView {
           .alignSelf(.end)
           .marginTop(6)
           .maxHeight(17)
-          .layout(mode: .adjustWidth)
+          .width(100%)
       }
   }
   
@@ -199,108 +201,8 @@ extension String {
   }
 }
 
-
 extension Reactive where Base: StyledTextInputView {
   var text: Observable<String> {
     return base.textRelay.asObservable()
-  }
-}
-
-class UnderlinedTextView: UITextView {
-  
-  /// 밑줄 간격
-  private let underLineHeight: CGFloat = 40
-  /// 엔터 눌렀을 때 텍스트 간의 간격
-  private let enterSpacing: CGFloat = 21
-  /// 밑줄 최대 라인 수
-  private let numberOfMaxLines: Int = 4
-  
-  private let fontStyle: UIFont
-  private let foregroundColor: UIColor
-  private let underLineColor: CGColor
-  private let walwalColor: UIColor
-  
-  init(
-    font: UIFont,
-    textColor: UIColor,
-    underLineColor: UIColor,
-    tintColor: UIColor
-  ) {
-    self.fontStyle = font
-    self.foregroundColor = textColor
-    self.underLineColor = underLineColor.cgColor
-    self.walwalColor = tintColor
-    
-    super.init(frame: .zero, textContainer: nil)
-    self.font = font
-    self.tintColor = tintColor
-  }
-  
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-  
-  func attributeText() {
-    let paragraphStyle = NSMutableParagraphStyle()
-    paragraphStyle.lineSpacing = enterSpacing
-    
-    // 속성 텍스트 설정
-    let attributes: [NSAttributedString.Key: Any] = [
-      .paragraphStyle: paragraphStyle,
-      .font: fontStyle,
-      .foregroundColor: foregroundColor
-    ]
-    let attributedString = NSMutableAttributedString(string: text, attributes: attributes)
-    
-    // 왈왈
-    let walwalAttributes: [NSAttributedString.Key: Any] = [
-      .foregroundColor: walwalColor
-    ]
-    if let range = text.range(of: "왈왈") {
-      let nsRange = NSRange(range, in: text)
-      attributedString.addAttributes(walwalAttributes, range: nsRange)
-    }
-    
-    if let range = text.range(of: "WalWal") {
-      let nsRange = NSRange(range, in: text)
-      attributedString.addAttributes(walwalAttributes, range: nsRange)
-    }
-    
-    // UITextView에 속성 텍스트 할당
-    self.attributedText = attributedString
-    self.typingAttributes[.paragraphStyle] = paragraphStyle
-  }
-  
-  override func draw(_ rect: CGRect) {
-    super.draw(rect)
-    
-    guard let context = UIGraphicsGetCurrentContext() else { return }
-    context.setStrokeColor(underLineColor)
-    context.setLineWidth(1.0)
-    
-    let lineSpacing: CGFloat = 0
-    let fontLineHeight: CGFloat = self.font?.lineHeight ?? 0
-    
-    var currentBaseline = self.textContainerInset.top + self.textContainer.lineFragmentPadding + underLineHeight - fontLineHeight - 10
-    
-    
-    
-    for _ in 0..<numberOfMaxLines {
-      let underlineY = currentBaseline + fontLineHeight
-      context.move(
-        to: CGPoint(
-          x: self.textContainerInset.left,
-          y: underlineY
-        )
-      )
-      context.addLine(
-        to: CGPoint(
-          x: rect.width - self.textContainerInset.right,
-          y: underlineY
-        )
-      )
-      context.strokePath()
-      currentBaseline += underLineHeight + lineSpacing
-    }
   }
 }
